+++ CHANGELOG.md
@@ -13,6 +13,7 @@ This package follows standard semvar, `<major>.<minor>.<build>`. No breaking cha
 * `hide-components` attribute to disable showing the components.
 * Fix components to actually collapse when `collapse` is set.
 * Fix displaying tag descriptions correctly and add supporting slot for subsections.
+* Fix schema expanded description text display
 
 ## 0.8 ##
 * Removing the `RESET` button as it is confusing for users to see. It only repopulated the defaults, and to do that, you can easily switch tabs and come back.